+++ client/src/pages/Search/searchcard.js
@@ -1,0 +1,37 @@
+import React from "react";
+import Card from 'react-bootstrap/card';
+import Form from 'react-bootstrap/form';
+import FormControl from 'react-bootstrap/formcontrol';
+import Button from 'react-bootstrap/button';
+import ListGroup from 'react-bootstrap/listgroup';
+
+function SearchCard() {
+  return (
+    <Card style={{ width: "18rem" }}>
+      <Form inline>
+        <FormControl
+          type="text"
+          placeholder='List Items'
+          className="mr-sm-2"
+        />
+        <Button variant="outline-success">
+          Search
+        </Button>
+      </Form>
+      <Card.Header>Featured</Card.Header>
+      <ListGroup variant="flush">
+        <ListGroup.Item>
+          Cras justo odio
+        </ListGroup.Item>
+        <ListGroup.Item>
+          Dapibus ac facilisis in
+        </ListGroup.Item>
+        <ListGroup.Item>
+          Vestibulum at eros
+        </ListGroup.Item>
+      </ListGroup>
+    </Card>
+  );
+}
+
+export default SearchCard;
